--- Dockerfile.dev
@@ -1,11 +1,0 @@
-FROM node:current-alpine as base
-
-WORKDIR /app
-
-COPY package*.json ./
-
-RUN npm ci
-
-COPY . .
-
-CMD ["npm", "run", "dev"]
