--- compose-dev.yaml
@@ -1,13 +1,0 @@
-services:
-  zerowave-portofolio:
-    image: zerowave-website
-    build: 
-      context: .
-      dockerfile : Dockerfile.dev
-    volumes: 
-      - .:/app
-      - /app/node_modules
-    ports: 
-      - "80:3000"
-
-
